requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
fastapi[standard]
pyjwt
datetime
quixstreams
werkzeug
asyncpg